Fixes #25160 - Make command runner a module in a compatible way
